% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fishhist.R
\docType{data}
\name{fishhist}
\alias{fishhist}
\title{The Southeast Fisheries Management History dataset}
\format{
A data frame
}
\usage{
data(fishhist)
}
\description{
A processed and cleaned management history dataframe
}
\keyword{datasets}
