% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mtypes.R
\name{plot_mtypes}
\alias{plot_mtypes}
\title{Percent of regulations by management type and sector}
\usage{
plot_mtypes(data = fishhist, FMP_filter = NULL)
}
\arguments{
\item{data}{default=fishhist}

\item{FMP}{Optional filter to plot only a single FMP of interest. If Null, the function will return a list containing a plot for each FMP.
\subsection{FMP:} {
   \describe{
     \item{ATLANTIC COAST RED DRUM FISHERY}
     \item{COASTAL MIGRATORY PELAGIC RESOURCES}
     \item{DOLPHIN AND WAHOO FISHERY OFF THE ATLANTIC STATES}
     \item{FMP FOR THE EEZ AROUND PUERTO RICO}
     \item{FMP FOR THE EEZ AROUND ST. CROIX}
     \item{FMP FOR THE EEZ AROUND ST. THOMAS AND ST. JOHN}
     \item{GOLDEN CRAB FISHERY OF THE SOUTH ATLANTIC}
     \item{QUEEN CONCH RESOURCES OF PUERTO RICO AND THE U.S. VIRGIN ISLANDS}
     \item{RED DRUM FISHERY OF THE GULF OF MEXICO}
     \item{REEF FISH FISHERY OF PUERTO RICO AND THE U.S. VIRGIN ISLANDS}
     \item{REEF FISH RESOURCES OF THE GULF OF MEXICO}
     \item{SHRIMP FISHERY OF THE GULF OF MEXICO}
     \item{SHRIMP FISHERY OF THE SOUTH ATLANTIC}
     \item{SNAPPER-GROUPER FISHERY OF THE SOUTH ATLANTIC REGION}
     \item{SPINY LOBSTER FISHERY OF PUERTO RICO AND THE U.S. VIRGIN ISLANDS}
     \item{SPINY LOBSTER FISHERY OF THE GULF OF MEXICO AND SOUTH ATLANTIC}
   }
}}
}
\value{
A plot or list of plots (e.g. use plot$plot[[1]] to view the first plot)
}
\description{
Percent of regulations by management type and sector
}
\examples{
plot_mtypes(fishhist, FMP_filter="REEF FISH RESOURCES OF THE GULF OF MEXICO")
}
